feat: Report how many counts the metric instance received
Instead of reporting the max and min of each `increment_counter`
request. Gives more information in the common case of sending counts
of 1 at the cost of not reporting the min and max value passed to
`increment_counter`.

In a perfectly orthogonal world a counter would not let us observe by
how much it was incremented with each time, but as long as we do we
should choose whatever is most useful. This makes increments observable
across metric collectors but not within them which should be more useful
since gauges can cover much of ground when wanting to observe min and
max in the same metric collector.