feat: Separate initialization of metrics_cloudwatch from driving it
Currently it isn't possible to wait for set_global_recorder to be called before proceeding, making it possible to drop metrics that are sent between the call to `init` and `set_global_recorder` being called.

By separating the driver task like this one can call init first, waiting to send and metrics until it is done, and then spawn the future on a tokio task.